+++ client/src/Components/LandingPage/Carousel.js
@@ -1,0 +1,37 @@
+import React from "react";
+import Slider from "react-slick";
+
+export class Carousel extends React.Component {
+    render() {
+      var settings = {
+        dots: true,
+        infinite: true,
+        speed: 500,
+        arrows: true,
+        slidesToShow: 1,
+        slidesToScroll: 1
+      };
+      return (
+        <Slider {...settings}>
+          <div>
+            <h3>1</h3>
+          </div>
+          <div>
+            <h3>2</h3>
+          </div>
+          <div>
+            <h3>3</h3>
+          </div>
+          <div>
+            <h3>4</h3>
+          </div>
+          <div>
+            <h3>5</h3>
+          </div>
+          <div>
+            <h3>6</h3>
+          </div>
+        </Slider>
+      );
+    }
+  };
