+++ client/src/Components/LandingPage/DesktopContainer.jsx
@@ -1,0 +1,95 @@
+import PropTypes from "prop-types";
+import React, { Component } from "react";
+import {
+  Button,
+  Container,
+  Menu,
+  Responsive,
+  Segment,
+  Visibility
+} from "semantic-ui-react";
+import { Link } from "react-router-dom";
+import {HomePageHeader} from './Header';
+
+// Heads up!
+// We using React Static to prerender our docs with server side rendering, this is a quite simple solution.
+// For more advanced usage please check Responsive docs under the "Usage" section.
+
+const getWidth = () => {
+    const isSSR = typeof window === "undefined";
+  
+    return isSSR ? Responsive.onlyTablet.minWidth : window.innerWidth;
+  };
+  
+/* Heads up!
+* Neither Semantic UI nor Semantic UI React offer a responsive navbar, however, it can be implemented easily.
+* It can be more complicated, but you can create really flexible markup.
+*/
+
+  class DesktopContainer extends Component {
+    state = {};
+  
+    hideFixedMenu = () => this.setState({ fixed: false });
+    showFixedMenu = () => this.setState({ fixed: true });
+  
+    render() {
+      const { children } = this.props;
+      const { fixed } = this.state;
+  
+      return (
+        <Responsive getWidth={getWidth} minWidth={Responsive.onlyTablet.minWidth}>
+          <Visibility
+            once={false}
+            onBottomPassed={this.showFixedMenu}
+            onBottomPassedReverse={this.hideFixedMenu}
+          >
+            <Segment
+              inverted
+              textAlign="center"
+              style={{ minHeight: 700, padding: "1em 0em" }}
+              vertical
+            >
+              <Menu
+                fixed={fixed ? "top" : null}
+                inverted={!fixed}
+                pointing={!fixed}
+                secondary={!fixed}
+                size="large"
+              >
+                <Container>
+                  <Menu.Item as="a" active>
+                    Home
+                  </Menu.Item>
+                  <Menu.Item position="right">
+                    <Link to="/login">
+                      <Button as="a" inverted={!fixed}>
+                        Log in
+                      </Button>
+                    </Link>
+                    <Link to="/register">
+                      <Button
+                        as="a"
+                        inverted={!fixed}
+                        primary={fixed}
+                        style={{ marginLeft: "0.5em" }}
+                      >
+                        Sign Up
+                      </Button>
+                    </Link>
+                  </Menu.Item>
+                </Container>
+              </Menu>
+              <HomePageHeader />
+            </Segment>
+          </Visibility>
+          {children}
+        </Responsive>
+      );
+    }
+  }
+  
+  DesktopContainer.propTypes = {
+    children: PropTypes.node
+  };
+
+export default DesktopContainer;
