+++ client/src/Components/LandingPage/Footer.js
@@ -1,0 +1,43 @@
+import React from "react";
+import {
+  Container,
+  Grid,
+  Header,
+  List,
+  Segment
+} from "semantic-ui-react";
+
+export const Footer = () => (
+    <Segment inverted vertical style={{ padding: "5em 0em" }}>
+        <Container>
+          <Grid divided inverted stackable>
+            <Grid.Row>
+              <Grid.Column width={3}>
+                <Header inverted as="h4" content="About" />
+                <List link inverted>
+                  <List.Item as="a">Sitemap</List.Item>
+                  <List.Item as="a">Contact Us</List.Item>
+                  <List.Item as="a">Religious Ceremonies</List.Item>
+                  <List.Item as="a">Gazebo Plans</List.Item>
+                </List>
+              </Grid.Column>
+              <Grid.Column width={3}>
+                <Header inverted as="h4" content="Services" />
+                <List link inverted>
+                  <List.Item as="a">Do We Want Services?</List.Item>
+                  <List.Item as="a">Centaur Training</List.Item>
+                  <List.Item as="a">Sleight of Hand Magic</List.Item>
+                  <List.Item as="a">Favorite X-Men</List.Item>
+                </List>
+              </Grid.Column>
+              <Grid.Column width={7}>
+                <Header as="h4" inverted>
+                  roostr.io
+                </Header>
+                <p>It's easy. Just give us your money and use our application.</p>
+              </Grid.Column>
+            </Grid.Row>
+          </Grid>
+        </Container>
+      </Segment>
+  );
