+++ client/src/Components/LandingPage/MobileContainer.jsx
@@ -1,0 +1,95 @@
+import PropTypes from "prop-types";
+import React, { Component } from "react";
+import {
+  Button,
+  Container,
+  Menu,
+  Responsive,
+  Segment,
+  Sidebar,
+  Icon
+} from "semantic-ui-react";
+import {HomePageHeader} from './Header';
+
+// Heads up!
+// We using React Static to prerender our docs with server side rendering, this is a quite simple solution.
+// For more advanced usage please check Responsive docs under the "Usage" section.
+
+const getWidth = () => {
+    const isSSR = typeof window === "undefined";
+  
+    return isSSR ? Responsive.onlyTablet.minWidth : window.innerWidth;
+  };
+  
+/* Heads up!
+* Neither Semantic UI nor Semantic UI React offer a responsive navbar, however, it can be implemented easily.
+* It can be more complicated, but you can create really flexible markup.
+*/
+
+class MobileContainer extends Component {
+  state = {};
+
+  handleSidebarHide = () => this.setState({ sidebarOpened: false });
+
+  handleToggle = () => this.setState({ sidebarOpened: true });
+
+  render() {
+    const { children } = this.props;
+    const { sidebarOpened } = this.state;
+
+    return (
+      <Responsive
+        as={Sidebar.Pushable}
+        getWidth={getWidth}
+        maxWidth={Responsive.onlyMobile.maxWidth}
+      >
+        <Sidebar
+          as={Menu}
+          animation="push"
+          inverted
+          onHide={this.handleSidebarHide}
+          vertical
+          visible={sidebarOpened}
+        >
+          <Menu.Item as="a" active>
+            Home
+          </Menu.Item>
+        </Sidebar>
+
+        <Sidebar.Pusher dimmed={sidebarOpened}>
+          <Segment
+            inverted
+            textAlign="center"
+            style={{ minHeight: 350, padding: "1em 0em" }}
+            vertical
+          >
+            <Container>
+              <Menu inverted pointing secondary size="large">
+                <Menu.Item onClick={this.handleToggle}>
+                  <Icon name="sidebar" />
+                </Menu.Item>
+                <Menu.Item position="right">
+                  <Button as="a" inverted>
+                    Log in
+                  </Button>
+                  <Button as="a" inverted style={{ marginLeft: "0.5em" }}>
+                    Sign Up
+                  </Button>
+                </Menu.Item>
+              </Menu>
+            </Container>
+            <HomePageHeader mobile />
+          </Segment>
+
+          {children}
+        </Sidebar.Pusher>
+      </Responsive>
+    );
+  }
+}
+
+MobileContainer.propTypes = {
+  children: PropTypes.node
+};
+
+export default MobileContainer;
